Install packages when "vendor/" does not exist
The script that updates the database tries to execute the tests after
all the data is updated even when the packages are not installed.

This commit will install the packages with Composer when they are not
already installed. It will assume that because you don't have the
"vendor/" directory you probably didn't install the dependencies.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>